+++ package-tests/test_base.py
@@ -8,6 +8,14 @@ import os
 import pytest
 
 
+def test_to_boolean():
+    response_positive = ['Yes', '1', 'y', 'Y', 'True']
+    resoponse_negative = ['No', '0', 'n', 'N', 'False']
+    for i in range(len(response_positive)):
+        assert base.to_boolean(response_positive[i]) is True
+        assert base.to_boolean(resoponse_negative[i]) is False
+
+
 def test_check_klass_name():
     """Test validation of entered class names
     """