+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 3.3.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Adding plone-compile-resources script in the buildout and extending .gitignore file
+  [kakshay21]
 
 
 3.3.0 (2018-05-24)