+++ CHANGES.rst
@@ -8,6 +8,10 @@ Changelog
   [kakshay21]
 
 - Adding plone-compile-resources script in addon and extending .gitignore file
+- Add update_locale script in bin/ folder to update locales
+  [kakshay21]
+
+- Add plone-compile-resources script in addon and extending .gitignore file
   [kakshay21]
 
 - Add conditional tests in content_type