+++ CHANGES.rst
@@ -4,10 +4,13 @@ Changelog
 3.3.1 (unreleased)
 ------------------
 
+- Add update_locale script in bin/ folder to update locales
+  [kakshay21]
+
 - Add a custom to_boolean jinja filter
   [kakshay21]
 
-- Adding plone-compile-resources script in addon and extending .gitignore file
+- Add plone-compile-resources script in addon and extending .gitignore file
   [kakshay21]
 
 - Add conditional tests in content_type